requirements.txt
fastapi
uvicorn[standard]
sqlmodel
mysqlclient
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
